// types/ProductCategory.ts
export enum ProductCategory {
    ELECTRONICS = 'Électronique',
    CLOTHING = 'Vêtements',
    HOME_APPLIANCES = 'Appareils ménagers',
    BOOKS = 'Livres',
    SPORTS = 'Sports',
    FOOD = 'Alimentation',
    BEAUTY = 'Beauté',
    TOYS = 'Jouets',
    AUTOMOTIVE = 'Automobile',
    HEALTH = 'Santé',
    OTHER = 'Autres',
    FURNITURE = 'Mobilier',
  }